requirements.txt
streamlit
pdfplumber
groq
pandas
networkx
matplotlib
python-dotenv
